refactor(navigation): centralize scroll logic and simplify overlay behavior
- Create useScrollNavigation hook for centralized scroll management
- Simplify SectionNavigationOverlay - always visible, no hide/show logic
- Add scroll position save/restore with section-based navigation
- Remove complex coordinate tracking, use simple currentSection state
- Improve auto-scroll after domain checking with better timing
- Clean up debug logs and unused code
